Add files via upload
The Snake Game environment provides a structured grid where the snake can 
move in four directions: up, down, left, and right. At any given moment, the game 
state is determined by the position of the snake's head, the positions of the food 
items, and the length and configuration of the snake's body. The primary objective 
for the snake agent is to maximize its score by efficiently consuming food items 
while avoiding collisions that would terminate the game.
This Project made by 
- Shrouk Mohamed
- Aya Mohammed Elqadry 
-Amal Mohammed Elsaeed Eltelbany 
- Tassnem Abdelrahman Hamed Abdelwahab 
-Menna Ashraf Hanafy